Create a test of manually offsetting mesh elements
This approximately recreates the BRL-CAD logic for brute-force
offsetting using boolean operations on individual mesh elements.
Purpose is to try and identify performance losses as compared to speed
observed in Manifold v2.4.5.  Note that this code uses the new Manifold
APIs and would need to be backported to run with the older codes.